arsip tiny to integer

+++ database/migrations/2024_08_08_054357_create_arsips_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->tinyInteger('tw');
             $table->string('jenis');
-            $table->tinyInteger('jumlah');
+            $table->integer('jumlah');
             $table->timestamps();
         });
     }
